steps 3-5: build `-ffi` adaptor crate, add uniffi bindings, verify they work
Unfortunately, this doesn't work. Copying the ffi crate from the libsql
spike was straightforward enough, but it turns out that Surrealdb
in embedded mode requires that it is run in within a Tokio runtime.

This is strictly incompatible with Uniffi's model, in which rust
futures are run on the foreign language's runtime.